moved useStateReducer to hooks path
